Move all logic to postinstall
